Allow disk-based persistence to be completely disabled (#2228)
* Allow in-memory Loki and extent store

* Update config

* Plumb config

* Switch to newer buffer stream library

* Add the config in more places and update markdowns

* Use built-in stream

* Refactor table creation to a single factory

* Move in-memory persistence to an env var

* Add QueueTestServerFactory to centralize test server creation

* Properly handle offset and count in extents

Fix tests

* Add in-memory tests to all variants

* Fix two flaky tests

* Make random local files have some newline characters so diff tools are usable

* Use a larger buffer for the download, the test is very slow otherwise

* Add design spec

* Add logging to MemoryExtentStore

* Make queue GC logging consistent with blob GC logging

* Add VS Code option

* Improve setting description

* Don't include docs in the VS Code extension

* Fix merge

* Prevent SQL and in-memory persistence at the same time

* Add MemoryExtentChunkStore for shared in-memory size tracking

* Use a shared MemoryExtentChunkStore by default

* Align Table and Queue cleaning approach with Blob

This clean-up for in-memory queue extents  as well as fixes broken Table clean-up in VS Code

* Add extentMemoryLimit option per review feedback

* Update spec with comments and add note about slower memory release

* Reject --location along with --inMemoryPersistence

* Fail fast on the binary tests if the binaries don't exist

* Properly tag tests

* Polish READMEs

* Address code review comments

* Eliminate unnecessary diff noise

* Change extentMemoryLimit to use megabytes as the unit

* Address comments

* Update README

* Fix typo

+++ tests/exe.test.ts
@@ -36,6 +36,7 @@ import {
   overrideRequest,
   restoreBuildRequestOptions
 } from "./testutils";
+import { existsSync } from "fs";
 
 // server address used for testing. Note that Azurite.exe has
 // server address of http://127.0.0.1:10000 and so on by default
@@ -53,6 +54,14 @@ configLogger(false);
 // Azure Storage Connection String (using SAS or Key).
 const testLocalAzuriteInstance = true;
 
+const binaryPath = ".\\release\\azurite.exe"
+
+function throwOnMissingBinary() {
+  if (!existsSync(binaryPath)) {
+    throw new Error("The Windows binary does not exist. You must build it first using 'npm run build:exe'.")
+  }
+}
+
 describe("exe test", () => {
   const tableService = Azure.createTableService(
     createConnectionStringForTest(testLocalAzuriteInstance)
@@ -65,11 +74,15 @@ describe("exe test", () => {
 
   let childPid: number;
 
+  beforeEach(() => throwOnMissingBinary())
+
   before(async () => {
+    throwOnMissingBinary()
+
     overrideRequest(requestOverride, tableService);
     tableName = getUniqueName("table");
     const child = execFile(
-      ".\\release\\azurite.exe",
+      binaryPath,
       ["--blobPort 11000", "--queuePort 11001", "--tablePort 11002"],
       { cwd: process.cwd(), shell: true, env: {} }
     );
@@ -116,10 +129,14 @@ describe("exe test", () => {
     tableService.removeAllListeners();
 
     await find("name", "azurite.exe", true).then((list: any) => {
-      process.kill(list[0].pid);
+      if (list.length > 0) {
+        process.kill(list[0].pid);
+      }
     });
 
-    process.kill(childPid);
+    if (childPid) {
+      process.kill(childPid);
+    }
   });
 
   describe("table test", () => {
